feat: enable multiple entries per location
BREAKING CHANGE: changes the model definition of user entry => flush the database or apply according `makemigrations`

Adds additional model `LocationEntry` which just resembles a marker that has a one-to-many relationship to `UserEntry`s
Adding an additional entry redirects to link with location_entry_id as param in url. If this url-param is set, the entry will be applied to the existing model

Display in HTML is currently done by just adding the html strings.

TODO: cleanup code a bit, clearly divide individual entries