feat: allow users to input polygons, lines, or markers
Surprisingly little change was needed on the data side: the previous PointField is almost identical to the new GeoJSONField.

On the front-end, more changes were needed to handle the arbitrary input and to render arbitrary data onto the map.

the lat/lon coordinates were removed from the creation URL in favor of a POST request with GeoJSON data.

regression: custom icons are currently not possible.